+++ toolkit/template/scaffold/project/cpp/qt5/console/main.cpp
@@ -2,7 +2,9 @@
 #include <QDebug>
 
 int main(int argc, char *argv[]) {
+  {% if enable_event_loop %}
   QCoreApplication app(argc, argv);
+  {%- endif %}
 
   qDebug() << "Hello World";
   {% if enable_event_loop %}